+++ application/controllers/AdminController.php
@@ -26,13 +26,25 @@ class AdminController extends Zend_Controller_Action
 
         $this->view->key_ign = getenv('PREVARISC_PLUGIN_IGNKEY');
         $this->view->key_googlemap = getenv('PREVARISC_PLUGIN_GOOGLEMAPKEY');
-        $this->view->ldap_enabled = getenv('PREVARISC_LDAP_ENABLED');
         $this->view->dbname = getenv('PREVARISC_DB_DBNAME');
         $this->view->db_url = getenv('PREVARISC_DB_HOST').(getenv('PREVARISC_DB_PORT') ? ':'.getenv('PREVARISC_DB_PORT') : '');
         $this->view->api_enabled = getenv('PREVARISC_SECURITY_KEY') != "";
         $this->view->proxy_enabled = getenv('PREVARISC_PROXY_ENABLED');
         $this->view->third_party_plugins = implode(', ', explode(';', getenv('PREVARISC_THIRDPARTY_PLUGINS')));
 
+        if (getenv('PREVARISC_CAS_ENABLED')) {
+            $this->view->authentification = "CAS";
+        } else if (getenv('PREVARISC_NTLM_ENABLED')) {
+            $this->view->authentification = "NTLM";
+        } else if (getenv('PREVARISC_LDAP_ENABLED')) {
+            $this->view->authentification = sprintf("LDAP + BDD : %s:%d/%s", 
+                getenv("PREVARISC_LDAP_HOST"), 
+                getenv("PREVARISC_LDAP_PORT"),
+                getenv("PREVARISC_LDAP_BASEDN"));
+        } else {
+            $this->view->authentification = "BDD";
+        }
+        
         $this->view->cache_adapter = $cache_config['adapter'];
         $this->view->cache_url = $cache_config['host']. ($cache_config['port'] ? ':'.$cache_config['port'] : '');
         $this->view->cache_lifetime = $cache_config['lifetime'];